Create a separate layout for media items with no reply to info.
Also, separate `Timeline__Row` components from `TimelineView`, as it was getting too large.
