+++ features/messages/impl/src/main/kotlin/io/element/android/features/messages/impl/timeline/TimelineView.kt
@@ -17,32 +17,16 @@
 package io.element.android.features.messages.impl.timeline
 
 import androidx.compose.animation.animateContentSize
-import androidx.compose.foundation.background
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.interaction.MutableInteractionSource
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.BoxScope
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.offset
-import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.layout.widthIn
-import androidx.compose.foundation.layout.wrapContentHeight
-import androidx.compose.foundation.layout.wrapContentSize
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.LazyListState
 import androidx.compose.foundation.lazy.itemsIndexed
 import androidx.compose.foundation.lazy.rememberLazyListState
 import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.ArrowDownward
 import androidx.compose.material3.MaterialTheme
@@ -57,53 +41,25 @@ import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.runtime.snapshotFlow
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.layout.LastBaseline
 import androidx.compose.ui.res.pluralStringResource
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.tooling.preview.PreviewParameter
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.zIndex
 import io.element.android.features.messages.impl.R
-import io.element.android.features.messages.impl.timeline.components.MessageEventBubble
-import io.element.android.features.messages.impl.timeline.components.MessageStateEventContainer
-import io.element.android.features.messages.impl.timeline.components.TimelineEventTimestampView
-import io.element.android.features.messages.impl.timeline.components.TimelineItemReactionsView
-import io.element.android.features.messages.impl.timeline.components.event.TimelineItemEventContentView
+import io.element.android.features.messages.impl.timeline.components.TimelineItemEventRow
+import io.element.android.features.messages.impl.timeline.components.TimelineItemStateEventRow
+import io.element.android.features.messages.impl.timeline.components.TimelineItemVirtualRow
 import io.element.android.features.messages.impl.timeline.components.group.GroupHeaderView
-import io.element.android.features.messages.impl.timeline.components.virtual.TimelineItemDaySeparatorView
-import io.element.android.features.messages.impl.timeline.components.virtual.TimelineLoadingMoreIndicator
 import io.element.android.features.messages.impl.timeline.model.TimelineItem
-import io.element.android.features.messages.impl.timeline.model.bubble.BubbleState
 import io.element.android.features.messages.impl.timeline.model.event.TimelineItemEventContent
 import io.element.android.features.messages.impl.timeline.model.event.TimelineItemEventContentProvider
-import io.element.android.features.messages.impl.timeline.model.event.TimelineItemImageContent
 import io.element.android.features.messages.impl.timeline.model.event.TimelineItemStateContent
-import io.element.android.features.messages.impl.timeline.model.event.TimelineItemVideoContent
-import io.element.android.features.messages.impl.timeline.model.virtual.TimelineItemDaySeparatorModel
-import io.element.android.features.messages.impl.timeline.model.virtual.TimelineItemLoadingModel
-import io.element.android.features.messages.impl.timeline.util.defaultTimelineContentPadding
-import io.element.android.libraries.designsystem.ElementTextStyles
-import io.element.android.libraries.designsystem.components.avatar.Avatar
-import io.element.android.libraries.designsystem.components.avatar.AvatarData
 import io.element.android.libraries.designsystem.preview.ElementPreviewDark
 import io.element.android.libraries.designsystem.preview.ElementPreviewLight
-import io.element.android.libraries.designsystem.theme.LocalColors
 import io.element.android.libraries.designsystem.theme.components.FloatingActionButton
 import io.element.android.libraries.designsystem.theme.components.Icon
-import io.element.android.libraries.designsystem.theme.components.Text
 import io.element.android.libraries.matrix.api.core.EventId
 import io.element.android.libraries.matrix.api.core.UserId
-import io.element.android.libraries.matrix.api.timeline.item.event.FileMessageType
-import io.element.android.libraries.matrix.api.timeline.item.event.ImageMessageType
-import io.element.android.libraries.matrix.api.timeline.item.event.InReplyTo
-import io.element.android.libraries.matrix.api.timeline.item.event.VideoMessageType
-import io.element.android.libraries.matrix.ui.components.AttachmentThumbnail
-import io.element.android.libraries.matrix.ui.components.AttachmentThumbnailInfo
-import io.element.android.libraries.matrix.ui.components.AttachmentThumbnailType
 import kotlinx.collections.immutable.ImmutableList
 import kotlinx.coroutines.flow.distinctUntilChanged
 import kotlinx.coroutines.launch
@@ -252,291 +208,6 @@ fun TimelineItemRow(
 }
 
 @Composable
-fun TimelineItemVirtualRow(
-    virtual: TimelineItem.Virtual,
-    modifier: Modifier = Modifier
-) {
-    when (virtual.model) {
-        is TimelineItemLoadingModel -> TimelineLoadingMoreIndicator(modifier)
-        is TimelineItemDaySeparatorModel -> TimelineItemDaySeparatorView(virtual.model, modifier)
-        else -> return
-    }
-}
-
-@Composable
-fun TimelineItemEventRow(
-    event: TimelineItem.Event,
-    isHighlighted: Boolean,
-    onClick: () -> Unit,
-    onLongClick: () -> Unit,
-    onUserDataClick: (UserId) -> Unit,
-    inReplyToClick: (EventId) -> Unit,
-    modifier: Modifier = Modifier
-) {
-    val interactionSource = remember { MutableInteractionSource() }
-
-    fun onUserDataClicked() {
-        onUserDataClick(event.senderId)
-    }
-
-    fun inReplayToClicked() {
-        val inReplyToEventId = (event.inReplyTo as? InReplyTo.Ready)?.eventId ?: return
-        inReplyToClick(inReplyToEventId)
-    }
-
-    val (parentAlignment, contentAlignment) = if (event.isMine) {
-        Pair(Alignment.CenterEnd, Alignment.End)
-    } else {
-        Pair(Alignment.CenterStart, Alignment.Start)
-    }
-
-    Box(
-        modifier = modifier
-            .fillMaxWidth()
-            .wrapContentHeight(),
-        contentAlignment = parentAlignment
-    ) {
-        Row {
-            if (!event.isMine) {
-                Spacer(modifier = Modifier.width(4.dp))
-            }
-            Column(horizontalAlignment = contentAlignment) {
-                if (event.showSenderInformation) {
-                    MessageSenderInformation(
-                        event.safeSenderName,
-                        event.senderAvatar,
-                        Modifier
-                            .zIndex(1f)
-                            .offset(y = 12.dp)
-                            .clickable(onClick = ::onUserDataClicked)
-                    )
-                }
-                val bubbleState = BubbleState(
-                    groupPosition = event.groupPosition,
-                    isMine = event.isMine,
-                    isHighlighted = isHighlighted,
-                )
-                MessageEventBubble(
-                    state = bubbleState,
-                    interactionSource = interactionSource,
-                    onClick = onClick,
-                    onLongClick = onLongClick,
-                    modifier = Modifier
-                        .zIndex(-1f)
-                        .widthIn(max = 320.dp)
-                ) {
-                    MessageEventBubbleContent(
-                        event = event,
-                        interactionSource = interactionSource,
-                        onMessageClick = onClick,
-                        onMessageLongClick = onLongClick,
-                        inReplyToClick = ::inReplayToClicked,
-                    )
-                }
-                TimelineItemReactionsView(
-                    reactionsState = event.reactionsState,
-                    modifier = Modifier
-                        .zIndex(1f)
-                        .offset(x = if (event.isMine) 0.dp else 20.dp, y = -(4.dp))
-                )
-            }
-            if (event.isMine) {
-                Spacer(modifier = Modifier.width(16.dp))
-            }
-        }
-    }
-    if (event.groupPosition.isNew()) {
-        Spacer(modifier = modifier.height(8.dp))
-    } else {
-        Spacer(modifier = modifier.height(2.dp))
-    }
-}
-
-@Composable
-fun TimelineItemStateEventRow(
-    event: TimelineItem.Event,
-    isHighlighted: Boolean,
-    onClick: () -> Unit,
-    onLongClick: () -> Unit,
-    modifier: Modifier = Modifier
-) {
-    val interactionSource = remember { MutableInteractionSource() }
-    Box(
-        modifier = modifier
-            .fillMaxWidth()
-            .wrapContentHeight(),
-        contentAlignment = Alignment.Center
-    ) {
-        MessageStateEventContainer(
-            isHighlighted = isHighlighted,
-            interactionSource = interactionSource,
-            onClick = onClick,
-            onLongClick = onLongClick,
-            modifier = Modifier
-                .zIndex(-1f)
-                .widthIn(max = 320.dp)
-        ) {
-            TimelineItemEventContentView(
-                content = event.content,
-                interactionSource = interactionSource,
-                onClick = onClick,
-                onLongClick = onLongClick,
-                modifier = Modifier.defaultTimelineContentPadding()
-            )
-        }
-    }
-}
-
-@Composable
-fun MessageEventBubbleContent(
-    event: TimelineItem.Event,
-    interactionSource: MutableInteractionSource,
-    onMessageClick: () -> Unit,
-    onMessageLongClick: () -> Unit,
-    inReplyToClick: () -> Unit,
-    modifier: Modifier = Modifier
-) {
-    val showTimestampWithOverlay = event.content is TimelineItemImageContent || event.content is TimelineItemVideoContent
-
-    @Composable
-    fun ContentView(
-        modifier: Modifier = Modifier
-    ) {
-        TimelineItemEventContentView(
-            content = event.content,
-            interactionSource = interactionSource,
-            onClick = onMessageClick,
-            onLongClick = onMessageLongClick,
-            modifier = modifier,
-        )
-    }
-
-    if (showTimestampWithOverlay) {
-        Box(modifier.wrapContentSize()) {
-            Column {
-                if (event.inReplyTo is InReplyTo.Ready) {
-                    val senderName = event.senderDisplayName ?: event.senderId.value
-                    val attachmentThumbnailInfo = attachmentThumbnailInfoForInReplyTo(event.inReplyTo)
-                    ReplyToContent(
-                        senderName = senderName,
-                        text = event.inReplyTo.content.body,
-                        attachmentThumbnailInfo = attachmentThumbnailInfo,
-                        modifier = Modifier.clickable(enabled = true, onClick = inReplyToClick),
-                    )
-                }
-                ContentView()
-            }
-            Box(
-                modifier = Modifier
-                    .padding(horizontal = 4.dp, vertical = 4.dp)
-                    .background(LocalColors.current.gray300, RoundedCornerShape(10.0.dp))
-                    .align(Alignment.BottomEnd)
-            ) {
-                TimelineEventTimestampView(
-                    event = event,
-                    onClick = onMessageClick,
-                    modifier = Modifier.padding(horizontal = 4.dp, vertical = 2.dp)
-                )
-            }
-        }
-    } else {
-        Column {
-            if (event.inReplyTo is InReplyTo.Ready) {
-                val senderName = event.senderDisplayName.takeIf { !it.isNullOrBlank() } ?: event.senderId.value
-                val attachmentThumbnailInfo = attachmentThumbnailInfoForInReplyTo(event.inReplyTo)
-                ReplyToContent(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .padding(start = 8.dp, end = 8.dp, top = 8.dp)
-                        .clickable(enabled = true, onClick = inReplyToClick),
-                    senderName = senderName,
-                    text = event.inReplyTo.content.body,
-                    attachmentThumbnailInfo = attachmentThumbnailInfo,
-                )
-            }
-            ContentView(modifier = Modifier.padding(start = 12.dp, end = 12.dp, top = 8.dp))
-            TimelineEventTimestampView(
-                event = event,
-                onClick = onMessageClick,
-                modifier = Modifier
-                    .align(Alignment.End)
-                    .padding(horizontal = 8.dp, vertical = 2.dp)
-            )
-        }
-    }
-}
-
-@Composable
-private fun ReplyToContent(
-    senderName: String,
-    text: String?,
-    attachmentThumbnailInfo: AttachmentThumbnailInfo?,
-    modifier: Modifier = Modifier,
-) {
-    val paddings = if (attachmentThumbnailInfo != null) {
-        PaddingValues(start = 4.dp, end = 12.dp, top = 4.dp, bottom = 4.dp)
-    } else {
-        PaddingValues(start = 12.dp, end = 12.dp, top = 8.dp, bottom = 4.dp)
-    }
-    Row(
-        modifier
-            .clip(RoundedCornerShape(6.dp))
-            .background(MaterialTheme.colorScheme.surface)
-            .padding(paddings)
-    ) {
-        if (attachmentThumbnailInfo != null) {
-            AttachmentThumbnail(
-                info = attachmentThumbnailInfo,
-                backgroundColor = MaterialTheme.colorScheme.surfaceVariant,
-                modifier = Modifier
-                    .size(36.dp)
-                    .clip(RoundedCornerShape(4.dp))
-            )
-            Spacer(modifier = Modifier.width(8.dp))
-        }
-        Column(verticalArrangement = Arrangement.SpaceBetween) {
-            Text(
-                senderName,
-                style = ElementTextStyles.Regular.caption2.copy(fontWeight = FontWeight.Medium),
-                textAlign = TextAlign.Start,
-                color = MaterialTheme.colorScheme.primary,
-                maxLines = 1,
-                overflow = TextOverflow.Ellipsis,
-            )
-
-            Text(
-                text = text.orEmpty(),
-                style = ElementTextStyles.Regular.caption1,
-                textAlign = TextAlign.Start,
-                color = LocalColors.current.placeholder,
-                maxLines = 1,
-                overflow = TextOverflow.Ellipsis,
-            )
-        }
-    }
-}
-
-@Composable
-private fun MessageSenderInformation(
-    sender: String,
-    senderAvatar: AvatarData?,
-    modifier: Modifier = Modifier
-) {
-    Row(modifier = modifier) {
-        if (senderAvatar != null) {
-            Avatar(senderAvatar)
-            Spacer(modifier = Modifier.width(4.dp))
-        }
-        Text(
-            text = sender,
-            style = MaterialTheme.typography.titleMedium,
-            modifier = Modifier
-                .alignBy(LastBaseline)
-        )
-    }
-}
-
-@Composable
 internal fun BoxScope.TimelineScrollHelper(
     lazyListState: LazyListState,
     timelineItems: ImmutableList<TimelineItem>,
@@ -595,29 +266,6 @@ internal fun BoxScope.TimelineScrollHelper(
         }
     }
 }
-
-private fun attachmentThumbnailInfoForInReplyTo(inReplyTo: InReplyTo.Ready) =
-    when (val type = inReplyTo.content.type) {
-        is ImageMessageType -> AttachmentThumbnailInfo(
-            mediaSource = type.info?.thumbnailSource,
-            textContent = inReplyTo.content.body,
-            type = AttachmentThumbnailType.Image,
-            blurHash = type.info?.blurhash,
-        )
-        is VideoMessageType -> AttachmentThumbnailInfo(
-            mediaSource = type.info?.thumbnailSource,
-            textContent = inReplyTo.content.body,
-            type = AttachmentThumbnailType.Video,
-            blurHash = type.info?.blurhash,
-        )
-        is FileMessageType -> AttachmentThumbnailInfo(
-            mediaSource = type.info?.thumbnailSource,
-            textContent = inReplyTo.content.body,
-            type = AttachmentThumbnailType.File,
-            blurHash = null,
-        )
-        else -> null
-    }
 
 @Preview
 @Composable
